+++ drools-yaml-rules-efesto/drools-yaml-rules-runtime/src/main/java/org/drools/yaml/runtime/RulesExecutorContainer.java
@@ -1,25 +1,20 @@
 package org.drools.yaml.runtime;
 
-import java.util.HashMap;
-import java.util.Map;
-
-import org.drools.yaml.runtime.service.JsonRulesExecutor;
-
 public enum RulesExecutorContainer {
 
-    INSTANCE;
-
-    private Map<Long, JsonRulesExecutor> rulesExecutors = new HashMap<>();
-
-    public void register(JsonRulesExecutor rulesExecutor) {
-        rulesExecutors.put(rulesExecutor.getId(), rulesExecutor);
-    }
-
-    public void dispose(JsonRulesExecutor rulesExecutor) {
-        rulesExecutors.remove(rulesExecutor.getId());
-    }
-
-    public JsonRulesExecutor get(Long id) {
-        return rulesExecutors.get(id);
-    }
+//    INSTANCE;
+//
+//    private Map<Long, JsonRulesExecutor> rulesExecutors = new HashMap<>();
+//
+//    public void register(JsonRulesExecutor rulesExecutor) {
+//        rulesExecutors.put(rulesExecutor.getId(), rulesExecutor);
+//    }
+//
+//    public void dispose(JsonRulesExecutor rulesExecutor) {
+//        rulesExecutors.remove(rulesExecutor.getId());
+//    }
+//
+//    public JsonRulesExecutor get(Long id) {
+//        return rulesExecutors.get(id);
+//    }
 }
